test(execution): rename test_execution_record_v2.py → test_execution_snapshot.py
Task 3 of H3. Follows the class + module rename: the dedicated
tests for the V2 class now live in test_execution_snapshot.py and
exercise ExecutionSnapshot. Four changes:

- git mv the old file to the new name
- update imports + constructor + classmethod references
- rename the three test functions with 'snapshot' in their names
- update the frozen test to catch Pydantic ValidationError (v2
  frozen models raise that, not AttributeError/TypeError)
- add a new test_execution_snapshot_model_dump verifying the
  .model_dump() API is available

Also updates test_execution_registry.py:
- test_list_executions_returns_dataclass →
  test_list_executions_returns_snapshot
- import + isinstance check use ExecutionSnapshot

Removes the temporary test file from Task 1.
